requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
pillow>=8.0.0
tqdm>=4.50.0
configparser>=5.0.0